stock-bull/data-pipeline/requirements.txt
pandas==2.1.0
numpy==1.24.3
yfinance==0.2.28
nsepy==0.8
newsapi-python==0.2.7
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
TA-Lib==0.4.28
schedule==1.2.0
transformers==4.30.0
torch==2.0.1